requirements.txt
anthropic>=0.33.0
python-dotenv>=1.0.1
pyautogui>=0.9.54
